Export attached files when export folder

+++ freedom/Action/Fdl/exportfld.php
@@ -3,7 +3,7 @@
  * Export Document from Folder
  *
  * @author Anakeen 2003
- * @version $Id: exportfld.php,v 1.21 2006/03/22 11:24:37 eric Exp $
+ * @version $Id: exportfld.php,v 1.22 2007/03/16 17:53:37 eric Exp $
  * @license http://opensource.org/licenses/gpl-license.php GNU Public License
  * @package FREEDOM
  * @subpackage 
@@ -24,15 +24,23 @@ function exportfld(&$action, $aflid="0", $famid="")
   $dbaccess = $action->GetParam("FREEDOM_DB");
   $fldid = GetHttpVars("id",$aflid);
   $wprof = (GetHttpVars("wprof","N")=="Y"); // with profil
+  $wfile = (GetHttpVars("wfile","N")=="Y"); // with files
   $fld = new_Doc($dbaccess, $fldid);
   if ($famid=="") $famid=GetHttpVars("famid");
   $tdoc = getChildDoc($dbaccess, $fldid,"0","ALL",array(),$action->user->id,"TABLE",$famid);
-    usort($tdoc,"orderbyfromid");
+  usort($tdoc,"orderbyfromid");
 
-  $foutname = uniqid("/tmp/exportfld").".csv";
+  $efldid=($fld->doctype=='D')?$fld->initid:"-";
+
+  if ($wfile) {
+    $foutdir=uniqid("/var/tmp/exportfld")."/";
+    if (! mkdir($foutdir)) exit();
+    
+    $foutname = $foutdir."fdl.csv";
+  } else {
+    $foutname = uniqid("/var/tmp/exportfld").".csv";
+  }
   $fout = fopen($foutname,"w");
-
-
 
   while (list($k,$doc)= each ($tdoc)) {        
     $docids[]=$doc->id;
@@ -52,6 +60,7 @@ function exportfld(&$action, $aflid="0", $famid="")
     $prevfromid = -1;
     reset($tdoc);
 
+    $ef=array(); //   files to export
     while (list($k,$zdoc)= each ($tdoc)) {
       $doc->Affect($zdoc,true);
 
@@ -59,7 +68,7 @@ function exportfld(&$action, $aflid="0", $famid="")
 	$adoc = $doc->getFamDoc();
 	if ($adoc->name != "") $fromname=$adoc->name;
 	else $fromname=$adoc->id;;
-	$lattr=$adoc->GetExportAttributes();
+	$lattr=$adoc->GetExportAttributes($wfile);
 	fputs($fout,"//FAM;".$adoc->title."(".$fromname.");<specid>;<fldid>;");
 	foreach($lattr as $ka=>$attr) {
 	  fputs($fout,str_replace(";"," - ",$attr->labelText).";");
@@ -75,19 +84,29 @@ function exportfld(&$action, $aflid="0", $famid="")
       reset($lattr);
       if ($doc->name != "") $name=$doc->name;
       else $name=$doc->id;
-      fputs($fout,"DOC;".$fromname.";".$name.";".$fldid.";");
+      fputs($fout,"DOC;".$fromname.";".$name.";".$efldid.";");
       // write values
-      while (list($ka,$attr)= each ($lattr)) {
+      foreach ($lattr as $ka=>$attr) {
       
 	$value= $doc->getValue($attr->id);
 	// invert HTML entities
+	if (($attr->type=="image") || ($attr->type=="file")) {
+	  $tfiles=$doc->vault_properties($attr);
+	  $tf=array();
+	  foreach ($tfiles as $f) {
+	    $fname=$f["vid"].'-'.$f["name"];
+	    $tf[]=$fname;
+	    $ef[$fname]=$f["path"];
+	  }
+	  $value=implode("\n",$tf);
+	} else {
 	$value = preg_replace("/(\&[a-zA-Z0-9\#]+;)/es", "strtr('\\1',\$trans)", $value);
  
 	// invert HTML entities which ascii code like &#232;
 
 	$value = preg_replace("/\&#([0-9]+);/es", "chr('\\1')", $value);
 
-	
+	}
 	fputs($fout,str_replace(array("\n",";","\r"),
 				array("\\n"," - ",""),
 				$value) .";");
@@ -140,9 +159,21 @@ function exportfld(&$action, $aflid="0", $famid="")
   fclose($fout);
 
   $fname=str_replace(array(" ","'"),array("_",""),$fld->title);
-  Http_DownloadFile($foutname, "$fname.csv", "text/csv");
-  unlink($foutname);
+  if ($wfile) {
+    foreach ($ef as $dest=>$source) {
+      if (!copy($source,$foutdir.$dest )) $err.=sprintf(_("cannot copy %s"),$dest);
+      
+    }
+    if ($err) $action->addWarningMsg($err);
+    system("cd $foutdir && zip fdl * > /dev/null",$ret);
+    $foutname=$foutdir."fdl.zip";
+    Http_DownloadFile($foutname, "$fname.zip", "application/x-zip",false,false);
+    
 
+  } else {
+    Http_DownloadFile($foutname, "$fname.csv", "text/csv",false,false);
+    unlink($foutname);
+  }
   exit;
 }
